Enable polymorphic ARN verification
When parsing an ARN, will send the parsed functor to every
arn_verify/2 to collect any errors they raise.

Unfortunately this is not compatible with modules, so this removes the
module declarations where necessary. Perhaps another method will make
this possible.

Adds an S3 Arn parser which is automatically called on any Arn for the
S3 service.

The pattern of collecting and returning errors lists is better than
printing an error and halting. For one thing it enables the ID of all
errors, not just the first. Also by turning errors into data, they can
be programmed.

issue #3